Move calibration module into pyFAI.gui.cli_calibration
