+++ pyFAI/app/mx_calibrate.py
@@ -43,7 +43,7 @@ __author__ = "Jerome Kieffer"
 __contact__ = "Jerome.Kieffer@ESRF.eu"
 __license__ = "MIT"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "20/02/2018"
+__date__ = "06/03/2018"
 __satus__ = "development"
 
 import logging
@@ -52,7 +52,7 @@ logging.basicConfig(level=logging.INFO)
 logging.captureWarnings(True)
 logger = logging.getLogger("pyFAI.mx_calibrate")
 
-import pyFAI.calibration
+from pyFAI.gui.cli_calibration import MultiCalib
 from pyFAI.third_party import six
 try:
     from rfoo.utils import rconsole
@@ -62,7 +62,7 @@ except ImportError:
 
 
 def main():
-    c = pyFAI.calibration.MultiCalib()
+    c = MultiCalib()
     c.parse()
     c.read_pixelsSize()
     c.read_dSpacingFile()
