+++ pyFAI/app/recalib.py
@@ -39,14 +39,14 @@ __author__ = "Jerome Kieffer"
 __contact__ = "Jerome.Kieffer@ESRF.eu"
 __license__ = "MIT"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "20/02/2018"
+__date__ = "06/03/2018"
 __satus__ = "development"
 
 import logging
 logging.basicConfig(level=logging.INFO)
 logging.captureWarnings(True)
 logger = logging.getLogger("pyFAI.recalib")
-from pyFAI.calibration import Recalibration
+from pyFAI.gui.cli_calibration import Recalibration
 from pyFAI.third_party import six
 try:
     from rfoo.utils import rconsole
